+++ BE/src/main/java/ifive/idrop/driver/service/DriverService.java
@@ -11,9 +11,10 @@ import ifive.idrop.common.exception.ErrorCode;
 import ifive.idrop.fcm.AlarmMessage;
 import ifive.idrop.fcm.NotificationUtill;
 import ifive.idrop.driver.repository.DriverRepository;
+import ifive.idrop.notification.domain.Notification;
 import ifive.idrop.parent.domain.Parent;
 import ifive.idrop.pickup.domain.PickUp;
-import ifive.idrop.repository.NotificationRepository;
+import ifive.idrop.notification.repository.NotificationRepository;
 import ifive.idrop.pickup.repository.PickUpRepository;
 import ifive.idrop.util.RequestSchedule;
 import lombok.RequiredArgsConstructor;
