+++ BE/src/main/java/ifive/idrop/notification/domain/Notification.java
@@ -1,4 +1,4 @@
-package ifive.idrop.entity;
+package ifive.idrop.notification.domain;
 
 import ifive.idrop.driver.domain.Driver;
 import jakarta.persistence.*;
